==> data/notebooks/2_Data_Ingestion_Transform.ipynb <==
# Replace these with your actual values:
storage_account_name = "<your storage account>" 
container_name = "<your container name>"
sas_token = "<your SAS Token>"

# Set the Spark configuration to use the SAS token
spark.conf.set(f"fs.azure.account.auth.type.{storage_account_name}.dfs.core.windows.net", "SAS")
spark.conf.set(f"fs.azure.sas.token.provider.type.{storage_account_name}.dfs.core.windows.net", "org.apache.hadoop.fs.azurebfs.sas.FixedSASTokenProvider")
spark.conf.set(f"fs.azure.sas.fixed.token.{storage_account_name}.dfs.core.windows.net", sas_token)

# Full path to the CSV in your raw container:
csv_file_path = f"abfss://{container_name}@{storage_account_name}.dfs.core.windows.net/sales_data.csv"

# Let's read it with Spark:
df = spark.read.option("header", "true").csv(csv_file_path)
display(df)

from pyspark.sql.functions import col

# Convert sales_qty to integer (if read as string)
df2 = df.withColumn("sales_qty", col("sales_qty").cast("int"))

# Example: filter out rows with sales_qty < 10
df_filtered = df2.filter(col("sales_qty") >= 10)

# Show results
df_filtered.show()

container_name = "<your 2nd container name>"


# Set the Spark configuration to use the SAS token
spark.conf.set(f"fs.azure.account.auth.type.{storage_account_name}.dfs.core.windows.net", "SAS")
spark.conf.set(f"fs.azure.sas.token.provider.type.{storage_account_name}.dfs.core.windows.net", 
               "org.apache.hadoop.fs.azurebfs.sas.FixedSASTokenProvider")
spark.conf.set(f"fs.azure.sas.fixed.token.{storage_account_name}.dfs.core.windows.net", sas_token)

parquet_path = f"abfss://{container_name}@{storage_account_name}.dfs.core.windows.net/curated_sales_data.parquet"
df_filtered.write.mode("overwrite").parquet(parquet_path)


# Ensure df_filtered is correctly defined and has data
if df_filtered is not None and df_filtered.count() > 0:
    df_filtered.write.mode("overwrite").parquet(parquet_path)
else:
    print("DataFrame is empty or not defined.")


parquet_path = "abfss://{contianer_name}@{storage_account_name}.dfs.core.windows.net/curated_sales_data.parquet/"

df_result = spark.read.parquet(parquet_path)
display(df_result)

dbutils.fs.ls("abfss://{container_name}@{storage_account_name}.dfs.core.windows.net/")

dbutils.fs.ls("abfss://{contianer_name}@{storage_account_name}.dfs.core.windows.net/")
